0x0c

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -24,7 +24,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 	p = malloc((strlen1 + n + 1) * sizeof(char));
 	if (p == NULL)
 		return (NULL);
-	for (i = 0, c = 0; i < (strlen1 + n); i++)
+	for (i = 0, c = 0; i < (strlen1 + (int) n); i++)
 	{
 		if (i < strlen1)
 			p[i] = s1[i];
@@ -35,4 +35,5 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 
 	return (p);
 }
+
